Move basegrid, xgrid, and uxgrid into _core subpackage
