+++ parcels/field.py
@@ -27,8 +27,8 @@ from parcels.converters import (
     UnitConverter,
     _unitconverters_map,
 )
-from parcels.uxgrid import UxGrid
-from parcels.xgrid import LEFT_OUT_OF_BOUNDS, RIGHT_OUT_OF_BOUNDS, XGrid, _transpose_xfield_data_to_tzyx
+from parcels._core.uxgrid import UxGrid
+from parcels._core.xgrid import LEFT_OUT_OF_BOUNDS, RIGHT_OUT_OF_BOUNDS, XGrid, _transpose_xfield_data_to_tzyx
 
 from ._index_search import GRID_SEARCH_ERROR, _search_time_index
