+++ parcels/interpolators.py
@@ -12,8 +12,8 @@ import parcels.utils.interpolation_utils as i_u
 
 if TYPE_CHECKING:
     from parcels.field import Field, VectorField
-    from parcels.uxgrid import _UXGRID_AXES
-    from parcels.xgrid import _XGRID_AXES
+    from parcels._core.uxgrid import _UXGRID_AXES
+    from parcels._core.xgrid import _XGRID_AXES
 
 __all__ = [
     "CGrid_Tracer",
