+++ tests/test_advection.py
@@ -27,7 +27,7 @@ from parcels.particle import Particle, Variable
 from parcels.particlefile import ParticleFile
 from parcels.particleset import ParticleSet
 from parcels.statuscodes import StatusCode
-from parcels.xgrid import XGrid
+from parcels import XGrid
 from tests.utils import round_and_hash_float_array
 
 kernel = {
