+++ tests/test_diffusion.py
@@ -11,7 +11,7 @@ from parcels.interpolators import XLinear
 from parcels.kernels import AdvectionDiffusionEM, AdvectionDiffusionM1, DiffusionUniformKh
 from parcels.particle import Particle, Variable
 from parcels.particleset import ParticleSet
-from parcels.xgrid import XGrid
+from parcels import XGrid
 from tests.utils import create_fieldset_zeros_conversion
