+++ tests/test_field.py
@@ -10,8 +10,8 @@ from parcels._datasets.structured.generic import T as T_structured
 from parcels._datasets.structured.generic import datasets as datasets_structured
 from parcels._datasets.unstructured.generic import datasets as datasets_unstructured
 from parcels.interpolators import UXPiecewiseConstantFace, UXPiecewiseLinearNode
-from parcels.uxgrid import UxGrid
-from parcels.xgrid import XGrid
+from parcels import UxGrid
+from parcels import XGrid
 
 
 def test_field_init_param_types():
