+++ tests/test_index_search.py
@@ -7,7 +7,7 @@ from parcels._datasets.structured.generic import datasets
 from parcels._index_search import _search_indices_curvilinear_2d
 from parcels._tutorial import download_example_dataset
 from parcels.field import Field
-from parcels.xgrid import XGrid
+from parcels import XGrid
 
 
 @pytest.fixture
