+++ tests/test_particleset.py
@@ -15,7 +15,7 @@ from parcels import (
     Variable,
 )
 from parcels._datasets.structured.generic import datasets as datasets_structured
-from parcels.xgrid import XGrid
+from parcels import XGrid
 from tests.common_kernels import DoNothing
 from tests.utils import round_and_hash_float_array
