+++ tests/test_spatialhash.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from parcels._datasets.structured.generic import datasets
-from parcels.xgrid import XGrid
+from parcels import XGrid
 
 
 def test_spatialhash_init():
